Add Quilt support (untested obviously)
This commit introduces a new system to the platform API: fallback platforms.
ExpectPlatform methods are queried first in the platform itself and then in
the fallback platforms (in order) if nothing is found.

This is used for Quilt as Architectury falls back to impl.fabric
if impl.quilt is not found.